refactor(*): move search into business layer
